Split plugin entrypoint test support
