+++ .github/ci-impact.json
@@ -468,7 +468,10 @@
         "tests/support/__init__.py",
         "tests/support/runtime_templates.py",
         "tests/support/runtime_instances.py",
-        "tests/support/plugin_entrypoints.py",
+        "tests/support/process_entrypoints.py",
+        "tests/support/search_entrypoints.py",
+        "tests/support/conjecture_entrypoints.py",
+        "tests/support/shrinking_entrypoints.py",
         "tests/support/plugin_noisy_module.py",
         "tests/support/services.py",
         "tests/support/state.py",
