+++ tests/boundary/storage/recovery/search_orchestration_support.py
@@ -13,10 +13,10 @@ def _install_search_plugin(
     runtime: JacobianRuntime,
     *,
     proposer_entrypoint: str = (
-        "tests.support.plugin_entrypoints:propose_fixture_values"
+        "tests.support.search_entrypoints:propose_fixture_values"
     ),
     refiner_entrypoint: str = (
-        "tests.support.plugin_entrypoints:refine_fixture_search"
+        "tests.support.search_entrypoints:refine_fixture_search"
     ),
     include_witness_oracle: bool = False,
 ) -> tuple[str, str]:
@@ -44,11 +44,11 @@ def _install_search_plugin(
     entrypoints = {
         "Proposer": proposer_entrypoint,
         "Refiner": refiner_entrypoint,
-        "Evaluator": "tests.support.plugin_entrypoints:evaluate_candidate",
+        "Evaluator": "tests.support.search_entrypoints:evaluate_candidate",
     }
     if include_witness_oracle:
         entrypoints["WitnessOracle"] = (
-            "tests.support.plugin_entrypoints:find_fixture_witness"
+            "tests.support.search_entrypoints:find_fixture_witness"
         )
     capabilities: dict[str, dict[str, str]] = {}
     for name, entrypoint in entrypoints.items():
